feat(kho): them so lo, don gia, han su dung vao KhoDialog va KhoList; fix Decimal JSON serialize; don_gia -> Integer

+++ BE/app/database/thuoc_vtyt.py
@@ -1,7 +1,6 @@
 from datetime import date
-from decimal import Decimal
 
-from sqlalchemy import Date, Integer, Numeric, String, Text
+from sqlalchemy import Date, Integer, String, Text
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.database.base import Base
@@ -21,7 +20,7 @@ class ThuocVtyt(Base):
     phan_loai: Mapped[str | None] = mapped_column(String(100), nullable=True)
     mo_ta: Mapped[str | None] = mapped_column(Text, nullable=True)
     han_su_dung: Mapped[date | None] = mapped_column(Date, nullable=True)
-    don_gia: Mapped[Decimal | None] = mapped_column(Numeric(15, 2), nullable=True)
+    don_gia: Mapped[int | None] = mapped_column(Integer, nullable=True)
     nha_san_xuat: Mapped[str | None] = mapped_column(String(255), nullable=True)
     hoat_chat: Mapped[str | None] = mapped_column(String(255), nullable=True)
     loai: Mapped[str | None] = mapped_column(String(10), nullable=True)
