+++ BE/app/schemas/thuoc_vtyt.py
@@ -1,5 +1,4 @@
 from datetime import date
-from decimal import Decimal
 
 from app.schemas.base import SchemaBase
 from pydantic import Field
@@ -15,7 +14,7 @@ class ThuocVtytBase(SchemaBase):
     phan_loai: str | None = Field(default=None, max_length=100)
     mo_ta: str | None = Field(default=None)
     han_su_dung: date | None = None
-    don_gia: Decimal | None = Field(default=None, ge=0)
+    don_gia: int | None = Field(default=None, ge=0)
     nha_san_xuat: str | None = Field(default=None, max_length=255)
     hoat_chat: str | None = Field(default=None, max_length=255)
     loai: str | None = Field(default=None, max_length=10)
@@ -33,7 +32,7 @@ class ThuocVtytUpdate(SchemaBase):
     phan_loai: str | None = Field(default=None, max_length=100)
     mo_ta: str | None = Field(default=None)
     han_su_dung: date | None = None
-    don_gia: Decimal | None = Field(default=None, ge=0)
+    don_gia: int | None = Field(default=None, ge=0)
     nha_san_xuat: str | None = Field(default=None, max_length=255)
     hoat_chat: str | None = Field(default=None, max_length=255)
     loai: str | None = Field(default=None, max_length=10)
